Change rule name; add docs
