+++ crates/ruff/src/codes.rs
@@ -727,7 +727,7 @@ pub fn code_to_rule(linter: Linter, code: &str) -> Option<Rule> {
         (Ruff, "007") => Rule::PairwiseOverZipped,
         (Ruff, "008") => Rule::MutableDataclassDefault,
         (Ruff, "009") => Rule::FunctionCallInDataclassDefaultArgument,
-        (Ruff, "010") => Rule::FStringConversion,
+        (Ruff, "010") => Rule::ExplicitFStringTypeConversion,
         (Ruff, "100") => Rule::UnusedNOQA,
 
         // flake8-django
